requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# APIs externas (SIN torch/CLIP)
pinecone-client==3.0.0
httpx==0.25.2

# Procesamiento básico
pillow==10.1.0
numpy==1.24.3

# Framework
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Para desarrollo
black==23.9.1